requirements.txt
fastapi
uvicorn
gunicorn
faiss-cpu
sentence-transformers
requests
pydantic
